feat(sidebar): one playlist list, not a server section beside it
The first piece of the redundancy this lot was aimed at. The sidebar had
a "Remote source" heading with the server's playlists under it, beside
the list holding the local ones — two lists of the same kind of thing,
which is exactly what unifying the navigation was supposed to end.

Server playlists now sit in the one playlist list, with the same chip the
library grid uses and a colour derived from their identifier, as the grid
and the remote view already did.

The filter deliberately does not reach the sidebar. It is navigation, not
a filtered view, and narrowing a tab to one source must not empty half of
the thing you navigate with.

That section also shipped hardcoded English — "No playlists on this
server yet", "N tracks", "not sent yet". A comment above it explained
that localisation was unnecessary because the feature was off by default
and no released build could reach it. The comment outlived the flag flip
that put sync_v2 in the default feature set nine days ago, and nothing
pointed that out because the section was the only surface those strings
had. The pending-creation subtext is now a key, and it carries exactly
the plural categories each locale already declares for the count phrasing
it extends.

The merge itself moved into a hook the sidebar and the library tab share,
so the two cannot drift on what a playlist row is. They pass different
local lists on purpose: the tab shows the user's own playlists, the
sidebar shows the smart ones too.
