+++ src/i18n/locales/en.json
@@ -207,7 +207,10 @@
     "myLibrary": {
       "playlistSubtext_zero": "0 tracks",
       "playlistSubtext_one": "{{count}} track",
-      "playlistSubtext_other": "{{count}} tracks"
+      "playlistSubtext_other": "{{count}} tracks",
+      "playlistSubtextPending_zero": "0 tracks · not sent yet",
+      "playlistSubtextPending_one": "{{count}} track · not sent yet",
+      "playlistSubtextPending_other": "{{count}} tracks · not sent yet"
     }
   },
   "topbar": {
