+++ src/i18n/locales/ja.json
@@ -207,7 +207,10 @@
     "myLibrary": {
       "playlistSubtext_zero": "0曲",
       "playlistSubtext_one": "{{count}}曲",
-      "playlistSubtext_other": "{{count}}曲"
+      "playlistSubtext_other": "{{count}}曲",
+      "playlistSubtextPending_zero": "0曲・未送信",
+      "playlistSubtextPending_one": "{{count}}曲・未送信",
+      "playlistSubtextPending_other": "{{count}}曲・未送信"
     }
   },
   "topbar": {
