+++ src/i18n/locales/ko.json
@@ -207,7 +207,10 @@
     "myLibrary": {
       "playlistSubtext_zero": "0곡",
       "playlistSubtext_one": "{{count}}곡",
-      "playlistSubtext_other": "{{count}}곡"
+      "playlistSubtext_other": "{{count}}곡",
+      "playlistSubtextPending_zero": "0곡 · 아직 보내지 않음",
+      "playlistSubtextPending_one": "{{count}}곡 · 아직 보내지 않음",
+      "playlistSubtextPending_other": "{{count}}곡 · 아직 보내지 않음"
     }
   },
   "topbar": {
