+++ src/i18n/locales/zh-TW.json
@@ -207,7 +207,10 @@
     "myLibrary": {
       "playlistSubtext_zero": "0 首曲目",
       "playlistSubtext_one": "{{count}} 曲目",
-      "playlistSubtext_other": "{{count}} 曲目"
+      "playlistSubtext_other": "{{count}} 曲目",
+      "playlistSubtextPending_zero": "0 首曲目 · 尚未傳送",
+      "playlistSubtextPending_one": "{{count}} 曲目 · 尚未傳送",
+      "playlistSubtextPending_other": "{{count}} 曲目 · 尚未傳送"
     }
   },
   "topbar": {
